split assignment operator implementation into copy and move cases.
I thought I was smart by using a trick, where the assignment operator swaps data with
a copy of the right argument. That however does not fully work, because there is no
way to check for object identity. (A copy is always created even if the code is something
like
`
    my_array = my_array;
`